pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chest-xray-classifier"
version = "1.2.0"
description = "Deep learning model for pneumonia detection in chest X-rays"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [{name = "Sameer Singh", email = "sameer@example.com"}]
keywords = ["pneumonia", "medical-imaging", "deep-learning", "classification"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Healthcare Industry",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "streamlit>=1.28.0",
    "pytorch-grad-cam>=1.4.5",
    "scikit-learn>=1.3.0",
    "pillow>=9.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]

[project.urls]
"Homepage" = "https://github.com/sameerfcb/Chest-X-Ray-Disease-Classifier"
"Bug Tracker" = "https://github.com/sameerfcb/Chest-X-Ray-Disease-Classifier/issues"
"Documentation" = "https://github.com/sameerfcb/Chest-X-Ray-Disease-Classifier/wiki"
"Source Code" = "https://github.com/sameerfcb/Chest-X-Ray-Disease-Classifier"

[project.scripts]
chest-xray-app = "src.app:main"

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | notebooks
)/
'''

[tool.isort]
profile = "black"
multi_line_mode = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 100
skip_glob = ["*/migrations/*"]

[tool.pylint.messages_control]
disable = [
    "C0103",  # Invalid name
    "C0111",  # Missing docstring
    "R0913",  # Too many arguments
    "R0914",  # Too many local variables
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --strict-markers"
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
